add enable_register_publish
Added enable_register_publish parameter to enable REGISTER and PUBLISH
method for topos
If set to 1 REGISTER and PUBLISH will be enabled for topos
Currently only via header will be applied for topos and if mask_callid
is set to 1 call-ID will be masked. Other headers should be manged by
the user as required.